Split data source file into files for each class in a data folder
