+++ torchtext/data/__init__.py
@@ -1,0 +1,17 @@
+from .batch import Batch
+from .dataset import Dataset, TabularDataset, ZipDataset
+from .example import Example
+from .field import Field, get_tokenizer
+from .iterator import (batch, BucketIterator, Iterator, BPTTIterator,
+                       pool, shuffled)
+from .pipeline import Pipeline
+from .utils import interleave_keys
+
+__all__ = ["Batch",
+           "Dataset", "TabularDataset", "ZipDataset",
+           "Example",
+           "Field", "get_tokenizer",
+           "batch", "BucketIterator", "Iterator", "BPTTIterator",
+           "pool", "shuffled",
+           "Pipeline",
+           "interleave_keys"]
