+++ torchtext/data/batch.py
@@ -1,0 +1,34 @@
+class Batch(object):
+    """Defines a batch of examples along with its Fields.
+
+    Attributes:
+        batch_size: Number of examples in the batch.
+        dataset: A reference to the dataset object the examples come from
+            (which itself contains the dataset's Field objects).
+        train: Whether the batch is from a training set.
+
+    Also stores the Variable for each column in the batch as an attribute.
+    """
+
+    def __init__(self, data=None, dataset=None, device=None, train=True):
+        """Create a Batch from a list of examples."""
+        if data is not None:
+            self.batch_size = len(data)
+            self.dataset = dataset
+            self.train = train
+            for (name, field) in dataset.fields.items():
+                if field is not None:
+                    setattr(self, name, field.numericalize(
+                        field.pad(x.__dict__[name] for x in data),
+                        device=device, train=train))
+
+    @classmethod
+    def fromvars(cls, dataset, batch_size, train=True, **kwargs):
+        """Create a Batch directly from a number of Variables."""
+        batch = cls()
+        batch.batch_size = batch_size
+        batch.dataset = dataset
+        batch.train = train
+        for k, v in kwargs.items():
+            setattr(batch, k, v)
+        return batch
