+++ torchtext/data/utils.py
@@ -1,0 +1,10 @@
+def interleave_keys(a, b):
+    """Interleave bits from two sort keys to form a joint sort key.
+
+    Examples that are similar in both of the provided keys will have similar
+    values for the key defined by this function. Useful for tasks with two
+    text fields like machine translation or natural language inference.
+    """
+    def interleave(args):
+        return ''.join([x for t in zip(*args) for x in t])
+    return int(''.join(interleave(format(x, '016b') for x in (a, b))), base=2)
